#21: Refactor API router and provider configurations
