+++ pkg/api/providers/google/index.go
@@ -1,0 +1,12 @@
+package google
+
+import (
+    "glide/pkg/api/providers"
+)
+
+func GoogleFullConfig() providers.ProviderConfigsAll {
+    return providers.ProviderConfigsAll {
+    Api: GoogleApiConfig,
+    Chat: GoogleChatDefaultConfig,
+    }
+}
